\name{Kpart.example}
\alias{Kpart.example}
\title{
An example of Kpart.
}
\description{
Shows example usage for the Kpart method.
}
\usage{
Kpart.example(data, K)
}
\arguments{
  \item{data}{ Two column matrix, the first column is respone, the second column is predictor in which to base knots. }
  \item{K}{ A positive interger that will partition the data K times. }
}
\details{
Example usage of Kpart package.
}
\value{
Produces a plot.
}
\references{
Tomas Fitzgerald 
}
\author{
Eric Golinko
}
\note{
thank you.
}

\examples{
# example()
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }
