\name{Kpart.plot}
\alias{Kpart.plot}
\title{
A Kpart plot
}
\description{
Displays plot of fitted values of lm.Kpart.
}
\usage{
Kpart.plot(data, K)
}
\arguments{
  \item{data}{  Two column matrix, the first column is respone, the second column is predictor in which to base knots.  }
  \item{K}{ A positive interger that will partition the data K times.  }
}
\details{
Plot of fitted values, as well as vertical lines representing the values of  knots selected.
}
\value{
Plot.
}
\references{
Tomas Fitzgerald
}
\author{
Eric Golinko
}
\note{
thank you.
}

\examples{
# example()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manpi }

